refacto(users): move id page api internal code
