+++ packages/~/app/urls/src/legacy.ts
@@ -8,8 +8,6 @@ type KnownPath =
   | "/"
   | "/legacy/moderations/:id"
   | "/legacy/users"
-  | "/legacy/users/:id"
-  | "/legacy/users/:id/reset"
   | "/moderations";
 
 //
