+++ packages/~/moderations/api/src/:id/page.tsx
@@ -84,6 +84,7 @@ export async function ModerationPage_Provider({
     var: { moncomptepro_pg },
   } = useRequestContext<MonComptePro_Pg_Context>();
 
+  // TODO(douglasduteil): return 404 status error
   if (!moderation_id) return <Moderation_NotFound />;
 
   const moderation = await moncomptepro_pg.query.moderations.findFirst({
@@ -94,6 +95,7 @@ export async function ModerationPage_Provider({
     },
   });
 
+  // TODO(douglasduteil): return 404 status error
   if (!moderation) return <Moderation_NotFound moderation_id={moderation_id} />;
 
   const users_organizations =
