+++ packages/~/users/api/package.json
@@ -13,6 +13,7 @@
     "@~/app.middleware": "workspace:*",
     "@~/app.urls": "workspace:*",
     "@~/organizations.repository": "workspace:*",
+    "@~/users.repository": "workspace:*",
     "hono": "4.0.7"
   },
   "devDependencies": {
