+++ packages/~/users/api/src/:id/index.tsx
@@ -1,0 +1,71 @@
+//
+
+import { zValidator } from "@hono/zod-validator";
+import type { Htmx_Header } from "@~/app.core/htmx";
+import { Entity_Schema } from "@~/app.core/schema";
+import { Main_Layout, userinfo_to_username } from "@~/app.layout/index";
+import type { Csp_Context } from "@~/app.middleware/csp_headers";
+import type { MonComptePro_Pg_Context } from "@~/app.middleware/moncomptepro_pg";
+import { hyyyyyypertool_session } from "@~/app.middleware/session";
+import type { UserInfo_Context } from "@~/app.middleware/vip_list.guard";
+import { api_ref } from "@~/app.urls/legacy";
+import { schema } from "@~/moncomptepro.database";
+import { eq } from "drizzle-orm";
+import { Hono } from "hono";
+import { jsxRenderer } from "hono/jsx-renderer";
+import user_organizations_page_route from "./organizations";
+import User_Page, { UserPage_Provider } from "./page";
+
+//
+
+export default new Hono<
+  MonComptePro_Pg_Context & UserInfo_Context & Csp_Context
+>()
+  .use("*", jsxRenderer(Main_Layout, { docType: true }))
+  .use("*", hyyyyyypertool_session)
+  .get(
+    "/",
+    zValidator("param", Entity_Schema),
+    async function GET({ req, render, var: { nonce, userinfo } }) {
+      const { id } = req.valid("param");
+
+      return render(
+        <UserPage_Provider value={{ id }}>
+          <User_Page />
+        </UserPage_Provider>,
+        {
+          nonce,
+          username: userinfo_to_username(userinfo),
+        },
+      );
+    },
+  )
+  .delete(
+    "/",
+    zValidator("param", Entity_Schema),
+    async function DELETE({ text, req, var: { moncomptepro_pg } }) {
+      const { id } = req.valid("param");
+      await moncomptepro_pg.delete(schema.users).where(eq(schema.users.id, id));
+      return text("OK", 200, {
+        "HX-Location": api_ref("/legacy/users", {}),
+      } as Htmx_Header);
+    },
+  )
+  .patch(
+    "/reset",
+    zValidator("param", Entity_Schema),
+    async function PATCH_RESET({ text, req, var: { moncomptepro_pg } }) {
+      const { id } = req.valid("param");
+      await moncomptepro_pg
+        .update(schema.users)
+        .set({
+          email_verified: false,
+        })
+        .where(eq(schema.users.id, id));
+      return text("OK", 200, { "HX-Refresh": "true" } as Htmx_Header);
+    },
+  )
+
+  .route("/organizations", user_organizations_page_route);
+
+//
