+++ packages/~/users/api/src/:id/organizations/index.tsx
@@ -2,14 +2,13 @@
 
 import { zValidator } from "@hono/zod-validator";
 import { Entity_Schema, Pagination_Schema } from "@~/app.core/schema";
-import type { MonComptePro_Pg_Context } from "@~/app.middleware/moncomptepro_pg";
 import { Hono } from "hono";
 import { jsxRenderer } from "hono/jsx-renderer";
 import { UserOrganizationTable, UserOrganizationTable_Provider } from "./Table";
 
 //
 
-export default new Hono<MonComptePro_Pg_Context>()
+export default new Hono()
   .use("/", jsxRenderer())
   .get(
     "/",
